new attributeEditor comp
